backend/slide-ingestion/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
Pillow==10.3.0
pydantic==2.7.1
aiofiles==23.2.1
python-multipart==0.0.9
